rename catalog_api to api
